requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.13.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
Authlib==1.3.2
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssselect2==0.7.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==33.1.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
kombu==5.4.2
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
multidict==6.1.0
oscrypto==1.3.0
passlib==1.7.4
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.0
PyMySQL==1.1.1
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
stripe==11.2.0
svglib==1.5.1
tinycss2==1.4.0
twilio==9.3.7
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
uvicorn[standard]
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.17.2
